requirements.txt
# coverage==7.4.0  # for testing
django==4.2.9
# gunicorn==21.2.0
# psycopg2-binary==2.9.9
# pytest-django==4.7.0  # for testing
# pytest==7.4.4  # for testing
asgiref==3.7.2
django-browser-reload==1.12.1
django-tailwind==3.8.0
pillow==10.2.0
sqlparse==0.4.4
# for requests
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
requests==2.31.0
urllib3==2.2.0
# huey
huey==2.5.0
# for thumbnails
django-imagekit==5.0.0
